数据库设计完成，且能正常写入数据库

+++ doc/sql/sql.sql
@@ -2,26 +2,26 @@
 -- user 用户表
 -- ---------
 CREATE TABLE user (
- user_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '用户ID',
- group_id mediumint(8) NOT NULL COMMENT '用户组ID',
+ user_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '用户ID',
+ group_id bigint(8) NOT NULL COMMENT '用户组ID',
  user_name varchar(32) NOT NULL COMMENT '用户名',
  user_pwd varchar(32) NOT NULL COMMENT '用户密码',
- user_phone int(12) NOT NULL COMMENT '用户手机号码',
+ user_phone bigint(12) NOT NULL COMMENT '用户手机号码',
  user_sex varchar(6) NOT NULL COMMENT '用户性别',
- user_qq mediumint(9) NOT NULL COMMENT '用户QQ号码',
+ user_qq bigint(9) NOT NULL COMMENT '用户QQ号码',
  user_email varchar(64) NOT NULL COMMENT '用户EMAIL地址',
  user_address varchar(255) NOT NULL COMMENT '用户地址',
- user_mark mediumint(9) NOT NULL COMMENT '用户积分',
+ user_mark bigint(9) NOT NULL COMMENT '用户积分',
  user_rank_id tinyint(3) NOT NULL COMMENT '用户等级',
  user_last_login_ip varchar(15) NOT NULL COMMENT '用户上一次登录IP地址',
 
- user_birthday int(13) NOT NULL COMMENT '用户生日',
+ user_birthday bigint(13) NOT NULL COMMENT '用户生日',
  user_description varchar(255) NOT NULL COMMENT '自我描述',
  user_image_url varchar(255) NOT NULL COMMENT '用户头像存储路径',
  user_school varchar(255) NOT NULL COMMENT '毕业学校',
- user_register_time int(13) NOT NULL COMMENT '用户注册时间',
+ user_register_time bigint(13) NOT NULL COMMENT '用户注册时间',
  user_register_ip varchar(15) NOT NULL COMMENT '用户注册时IP地址',
- user_last_update_time int(13) NOT NULL COMMENT '用户上次更新博客时间',
+ user_last_update_time bigint(13) NOT NULL COMMENT '用户上次更新博客时间',
  user_weibo varchar(255) NOT NULL COMMENT '用户微博',
  user_blood_type char(3) NOT NULL COMMENT '用户血型',
  user_says varchar(255) NOT NULL COMMENT '用户语录',
@@ -36,9 +36,9 @@ CREATE TABLE user (
 -- user_rank 用户权限表
 -- ---------------
 CREATE TABLE user_rank (
- rank_id mediumint(5) NOT NULL AUTO_INCREMENT COMMENT '自增ID',
+ rank_id bigint(5) NOT NULL AUTO_INCREMENT COMMENT '自增ID',
  user_rank_id smallint(5) NOT NULL COMMENT '等级ID',
- rank_mark mediumint(6) NOT NULL COMMENT '等级积分',
+ rank_mark bigint(6) NOT NULL COMMENT '等级积分',
  rank_name varchar(32) NOT NULL COMMENT '等级名称',
  PRIMARY KEY (rank_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
@@ -73,8 +73,8 @@ CREATE TABLE power_list (
 -- -----------------
 CREATE TABLE friend (
  f_id smallint(5) NOT NULL AUTO_INCREMENT COMMENT '自增ID',
- user_id mediumint(8) NOT NULL COMMENT '用户ID',
- friend_id mediumint(8) NOT NULL COMMENT '好友ID',
+ user_id bigint(8) NOT NULL COMMENT '用户ID',
+ friend_id bigint(8) NOT NULL COMMENT '好友ID',
  PRIMARY KEY (f_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
 
@@ -84,8 +84,8 @@ CREATE TABLE friend (
 -- --------------------
 CREATE TABLE user_attention (
  a_id smallint(5) NOT NULL AUTO_INCREMENT COMMENT '自增ID',
- user_id mediumint(8) NOT NULL COMMENT '用户ID',
- attention_id mediumint(8) NOT NULL COMMENT '关注ID',
+ user_id bigint(8) NOT NULL COMMENT '用户ID',
+ attention_id bigint(8) NOT NULL COMMENT '关注ID',
  PRIMARY KEY (a_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
 
@@ -94,9 +94,9 @@ CREATE TABLE user_attention (
 -- secret_message 用户私信表
 -- ---------------------
 CREATE TABLE secret_message (
- secret_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '自增私信ID',
- send_id mediumint(8) NOT NULL COMMENT '发信者ID',
- receive_id mediumint(8) NOT NULL COMMENT '收信者ID',
+ secret_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '自增私信ID',
+ send_id bigint(8) NOT NULL COMMENT '发信者ID',
+ receive_id bigint(8) NOT NULL COMMENT '收信者ID',
  message_topic varchar(64) NOT NULL COMMENT '私信标题',
  message_content varchar(255) NOT NULL COMMENT '私信内容',
  PRIMARY KEY (secret_id)
@@ -108,10 +108,10 @@ CREATE TABLE secret_message (
 -- system_message 系统通知表
 -- ----------------------
 CREATE TABLE system_message (
- system_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '系统通知ID',
- send_id mediumint(8) NOT NULL COMMENT '接受者ID',
+ system_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '系统通知ID',
+ send_id bigint(8) NOT NULL COMMENT '接受者ID',
  group_id tinyint(3) NOT NULL COMMENT '用户组ID',
- send_default mediumint(8) NOT NULL COMMENT '1时发送所有用户，0时则不采用',
+ send_default bigint(8) NOT NULL COMMENT '1时发送所有用户，0时则不采用',
  system_topic varchar(60) NOT NULL COMMENT '通知内容',
  system_content varchar(255) NOT NULL COMMENT '通知内容',
  PRIMARY KEY (system_id)
@@ -141,12 +141,12 @@ CREATE TABLE ad (
  ad_name varchar(60) NOT NULL COMMENT '该条广告记录的广告名称',
  ad_link varchar(255) NOT NULL COMMENT '广告链接地址',
  ad_code text NOT NULL COMMENT '广告链接的表现,文字广告就是文字或图片和flash就是它们的地址',
- start_time int(13) NOT NULL DEFAULT 0 COMMENT '广告开始时间',
- end_time int(13) NOT NULL DEFAULT 0 COMMENT '广告结束时间',
+ start_time bigint(13) NOT NULL DEFAULT 0 COMMENT '广告开始时间',
+ end_time bigint(13) NOT NULL DEFAULT 0 COMMENT '广告结束时间',
  link_man varchar(60) NOT NULL COMMENT '广告联系人',
  link_email varchar(60) NOT NULL COMMENT '广告联系人的邮箱',
  link_phone varchar(60) NOT NULL COMMENT '广告联系人得电话',
- click_count mediumint(8) NOT NULL DEFAULT 0 COMMENT '广告点击次数',
+ click_count bigint(8) NOT NULL DEFAULT 0 COMMENT '广告点击次数',
  enabled tinyint(3) NOT NULL DEFAULT 1 COMMENT '该广告是否关闭;1开启; 0关闭; 关闭后广告将不再有效',
  PRIMARY KEY (ad_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
@@ -157,11 +157,11 @@ CREATE TABLE ad (
 -- ------------------
 CREATE TABLE stay_message (
  stay_id smallint(5) NOT NULL AUTO_INCREMENT COMMENT '留言表自增ID',
- user_id mediumint(8) NOT NULL COMMENT '用户ID',
- stay_user_id mediumint(8) NOT NULL COMMENT '留言者ID',
+ user_id bigint(8) NOT NULL COMMENT '用户ID',
+ stay_user_id bigint(8) NOT NULL COMMENT '留言者ID',
  message_content varchar(255) NOT NULL COMMENT '留言内容',
  stay_user_ip varchar(15) NOT NULL COMMENT '留言用户的IP地址',
- message_stay_time int(13) NOT NULL COMMENT '留言时间',
+ message_stay_time bigint(13) NOT NULL COMMENT '留言时间',
  place varchar(64) NOT NULL COMMENT '地区',
  PRIMARY KEY (stay_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
@@ -171,7 +171,7 @@ CREATE TABLE stay_message (
 -- about_blog 博客信息表
 -- --------------------
 CREATE TABLE about_blog (
- blog_id mediumint(8) NOT NULL  COMMENT '用户ID',
+ blog_id bigint(8) NOT NULL  COMMENT '用户ID',
  blog_keyword varchar(255) NOT NULL COMMENT '博客关键字',
  blog_description varchar(255) NOT NULL COMMENT '博客描述',
  blog_name varchar(36) NOT NULL COMMENT '博客名称',
@@ -185,13 +185,13 @@ CREATE TABLE about_blog (
 -- visitor 最近访客表
 -- ----------------------
 CREATE TABLE visitor (
- v_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '访客记录ID',
- visitor_id mediumint(8) NOT NULL COMMENT '访客ID',
- visitor_time int(13) NOT NULL COMMENT '来访时间',
- user_id mediumint(8) NOT NULL COMMENT '被访用户ID',
+ v_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '访客记录ID',
+ visitor_id bigint(8) NOT NULL COMMENT '访客ID',
+ visitor_time bigint(13) NOT NULL COMMENT '来访时间',
+ user_id bigint(8) NOT NULL COMMENT '被访用户ID',
  visitor_ip varchar(15) NOT NULL COMMENT '访客IP地址',
  type_id int(3) NOT NULL COMMENT '访问板块ID',
- where_id mediumint(8) NOT NULL COMMENT '查看某板块的某个子项目，如查看相册板块的第3个相册，该ID对应该相册的ID号',
+ where_id bigint(8) NOT NULL COMMENT '查看某板块的某个子项目，如查看相册板块的第3个相册，该ID对应该相册的ID号',
  PRIMARY KEY (v_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
 
@@ -201,9 +201,9 @@ CREATE TABLE visitor (
 -- shuoshuo 用户心情说说表
 -- -----------------------
 CREATE TABLE shuoshuo (
- shuo_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '说说记录ID',
- user_id mediumint(8) NOT NULL COMMENT '用户ID',
- shuo_time int(13) NOT NULL DEFAULT 0 COMMENT '发布时间',
+ shuo_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '说说记录ID',
+ user_id bigint(8) NOT NULL COMMENT '用户ID',
+ shuo_time bigint(13) NOT NULL DEFAULT 0 COMMENT '发布时间',
  shuo_ip varchar(15) NOT NULL COMMENT '说说发布时的IP地址',
  shuoshuo varchar(255) NOT NULL COMMENT '说说内容',
  type_id tinyint(3) NOT NULL DEFAULT 3 COMMENT '栏目ID,默认为3',
@@ -216,15 +216,15 @@ CREATE TABLE shuoshuo (
 -- photo_sort 相片分类表
 -- -----------------------
 CREATE TABLE photo_sort (
- sort_img_id mediumint(8) NOT NULL  AUTO_INCREMENT COMMENT '相册ID',
+ sort_img_id bigint(8) NOT NULL  AUTO_INCREMENT COMMENT '相册ID',
  sort_img_name varchar(20) NOT NULL COMMENT '相册名',
  sort_img_type varchar(20) NOT NULL COMMENT '展示方式 0->仅主人可见,1->输入密码即可查看,2->仅好友能查看,3->回答问题即可查看',
  img_password varchar(32) NOT NULL COMMENT '查看密码',
- user_id mediumint(8) NOT NULL COMMENT '所属用户ID',
+ user_id bigint(8) NOT NULL COMMENT '所属用户ID',
  img_sort_question varchar(255) NOT NULL COMMENT '访问问题',
  img_sort_answer varchar(128) NOT NULL COMMENT '访问问题的答案',
  type_id int(3) NOT NULL DEFAULT 1 COMMENT '默认1表示相册板块',
- top_pic_src mediumint(8) NOT NULL COMMENT '封面图片的路径',
+ top_pic_src bigint(8) NOT NULL COMMENT '封面图片的路径',
  PRIMARY KEY (sort_img_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
 
@@ -234,13 +234,13 @@ CREATE TABLE photo_sort (
 -- photos 相片表
 -- -----------------------
 CREATE TABLE photos (
- photo_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '相片ID',
+ photo_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '相片ID',
  photo_name varchar(255) NOT NULL COMMENT '相片名称',
  photo_src varchar(255) NOT NULL COMMENT '图片路径',
  photo_description varchar(255) NOT NULL COMMENT '图片描述',
- user_id mediumint(8) NOT NULL COMMENT '所属用户ID',
- sort_id mediumint(8) NOT NULL COMMENT '所属相册ID',
- upload_time int(13) NOT NULL COMMENT '图片上传时间',
+ user_id bigint(8) NOT NULL COMMENT '所属用户ID',
+ sort_id bigint(8) NOT NULL COMMENT '所属相册ID',
+ upload_time bigint(13) NOT NULL COMMENT '图片上传时间',
  upload_ip varchar(15) NOT NULL COMMENT '图片操作上传IP地址',
  PRIMARY KEY (photo_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
@@ -252,8 +252,8 @@ CREATE TABLE photos (
 -- article_sort 文章分类表
 -- -------------------------
 CREATE TABLE article_sort (
- sort_article_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '文章自增ID',
- user_id mediumint(8) NOT NULL COMMENT '该分类所属用户',
+ sort_article_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '文章自增ID',
+ user_id bigint(8) NOT NULL COMMENT '该分类所属用户',
  sort_article_name varchar(60) NOT NULL COMMENT '分类名称',
  PRIMARY KEY (sort_article_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
@@ -266,11 +266,11 @@ CREATE TABLE article_sort (
 CREATE TABLE article (
  article_id smallint(5) NOT NULL AUTO_INCREMENT COMMENT '日志自增ID号',
  article_name varchar(128) NOT NULL COMMENT '文章名称',
- article_time int(13) NOT NULL COMMENT '发布时间',
+ article_time bigint(13) NOT NULL COMMENT '发布时间',
  article_ip varchar(15) NOT NULL COMMENT '发布IP',
  article_click int(10) NOT NULL COMMENT '查看人数',
- sort_article_id mediumint(8) NOT NULL COMMENT '所属分类',
- user_id mediumint(8) NOT NULL COMMENT '所属用户ID',
+ sort_article_id bigint(8) NOT NULL COMMENT '所属分类',
+ user_id bigint(8) NOT NULL COMMENT '所属用户ID',
  type_id tinyint(3) NOT NULL DEFAULT 1 COMMENT '栏目ID',
  article_type int(13) NOT NULL DEFAULT 1 COMMENT '文章的模式:0为私有，1为公开，2为仅好友查看',
  article_content text NOT NULL COMMENT '文章内容',
@@ -286,13 +286,13 @@ CREATE TABLE article (
 -- user_comment 用户评论表
 -- --------------------------
 CREATE TABLE user_comment (
- c_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '评论自增ID号',
- user_id mediumint(8) NOT NULL COMMENT '收到评论的用户ID',
+ c_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '评论自增ID号',
+ user_id bigint(8) NOT NULL COMMENT '收到评论的用户ID',
  type_id tinyint(3) NOT NULL COMMENT '评论栏目ID',
- commit_id mediumint(8) NOT NULL COMMENT '评论内容的ID',
+ commit_id bigint(8) NOT NULL COMMENT '评论内容的ID',
  commit_content varchar(255) NOT NULL COMMENT '评论内容',
- commit_user_id mediumint(8) NOT NULL COMMENT '评论者ID',
- commit_time int(13) NOT NULL COMMENT '评论时间',
+ commit_user_id bigint(8) NOT NULL COMMENT '评论者ID',
+ commit_time bigint(13) NOT NULL COMMENT '评论时间',
  commit_ip varchar(15) NOT NULL COMMENT '评论时的IP地址',
  PRIMARY KEY (c_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
@@ -303,10 +303,10 @@ CREATE TABLE user_comment (
 -- phone_message 短信记录表
 -- ----------------------------
 CREATE TABLE phone_message (
- phone_id mediumint(8) NOT NULL AUTO_INCREMENT COMMENT '自增ID号',
+ phone_id bigint(8) NOT NULL AUTO_INCREMENT COMMENT '自增ID号',
  phone_num varchar(12) NOT NULL COMMENT '用户手机号码',
  contents varchar(255) NOT NULL COMMENT '发送内容',
- send_time int(13) NOT NULL COMMENT '发送时间',
- user_id mediumint(8) NOT NULL COMMENT '用户ID',
+ send_time bigint(13) NOT NULL COMMENT '发送时间',
+ user_id bigint(8) NOT NULL COMMENT '用户ID',
  PRIMARY KEY (phone_id)
 ) ENGINE=INNODB DEFAULT CHARSET=utf8 ;
